deployment/requirements.txt
tensorflow==2.17.0
pillow==10.4.0
numpy==1.26.4